Fix tabplayer_gd compile errors
Agent-Logs-Url: https://github.com/goguitar/tabplayerV2/sessions/1cb8ace4-3a0c-46cb-8d00-832a66b8bf6b

+++ rust/tabplayer_gd/src/scenes/settings_page.rs
@@ -71,7 +71,7 @@ impl SettingsPage {
             let mut picker = ColorPickerButton::new_alloc();
             picker.set_pick_color(self.settings.string_colours[i]);
             picker.set_edit_alpha(false);
-            let args = [i.to_variant()];
+            let args = [(i as i64).to_variant()];
             let callable = self.base_mut().callable("on_color_changed").bind(&args);
             picker.connect("popup_closed", &callable);
             self.color_pickers.push(picker.clone());
